[sgx-untrusted-alloc] Introduce MaybeUntrusted and UntrustedBox
The new version introduces two concepts. The first one is `MaybeUntrusted`,
which is a marker trait for types that is safe to be used in untrusted
memory. The second one is `UntrustedBox<T>`, which is the untrusted
version of `Box<T>`, making allocating and access untrusted memory
easily.

With the two new concepts, all code that communicates with the host through
shared untrusted buffer is refactored to use `UntrustedBox`.